Refactor `factory.Executable()` to be a method rather than a func
This way, factory can satisfy an interface that requires `Executable()`.
